feat(toughts): cleaning folders and adding folder (#3)
