--- examples/schemas/product_sample.json
@@ -1,18 +1,0 @@
-{
-  "id": "PROD-123456",
-  "name": "Wireless Bluetooth Headphones",
-  "description": "Premium noise-cancelling wireless headphones with 30-hour battery life and comfortable over-ear design.",
-  "price": 149.99,
-  "category": "electronics",
-  "inventory": {
-    "quantity": 250,
-    "sku": "SKU-ABCD1234",
-    "warehouse": "East Coast Distribution Center"
-  },
-  "dimensions": {
-    "length": 7.5,
-    "width": 6.3,
-    "height": 3.2,
-    "weight": 0.65
-  }
-}
