--- examples/schemas/user_profile.json
@@ -1,143 +1,0 @@
-{
-  "id": "user_profile",
-  "name": "User Profile Schema",
-  "description": "Schema for validating user profile data",
-  "version": "1.0.0",
-  "fields": [
-    {
-      "id": "username",
-      "name": "Username",
-      "description": "The user's unique username",
-      "type": "string",
-      "required": true,
-      "validation": [
-        {
-          "type": "min_length",
-          "value": 3,
-          "message": "Username must be at least 3 characters long",
-          "error_code": "username_too_short"
-        },
-        {
-          "type": "max_length",
-          "value": 50,
-          "message": "Username cannot exceed 50 characters",
-          "error_code": "username_too_long"
-        },
-        {
-          "type": "pattern",
-          "value": "^[a-zA-Z0-9_]+$",
-          "message": "Username can only contain letters, numbers, and underscores",
-          "error_code": "username_invalid_chars"
-        }
-      ]
-    },
-    {
-      "id": "email",
-      "name": "Email Address",
-      "description": "The user's email address",
-      "type": "string",
-      "required": true,
-      "validation": [
-        {
-          "type": "pattern",
-          "value": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
-          "message": "Invalid email address format",
-          "error_code": "email_invalid_format"
-        }
-      ]
-    },
-    {
-      "id": "age",
-      "name": "Age",
-      "description": "The user's age in years",
-      "type": "integer",
-      "required": false,
-      "validation": [
-        {
-          "type": "min_value",
-          "value": 13,
-          "message": "User must be at least 13 years old",
-          "error_code": "age_too_young"
-        },
-        {
-          "type": "max_value",
-          "value": 120,
-          "message": "Age cannot exceed 120 years",
-          "error_code": "age_too_old"
-        }
-      ]
-    },
-    {
-      "id": "address",
-      "name": "Address",
-      "description": "The user's address information",
-      "type": "object",
-      "required": false,
-      "properties": [
-        {
-          "id": "street",
-          "name": "Street",
-          "type": "string",
-          "required": true
-        },
-        {
-          "id": "city",
-          "name": "City",
-          "type": "string",
-          "required": true
-        },
-        {
-          "id": "state",
-          "name": "State",
-          "type": "string",
-          "required": true
-        },
-        {
-          "id": "zip",
-          "name": "ZIP Code",
-          "type": "string",
-          "required": true,
-          "validation": [
-            {
-              "type": "pattern",
-              "value": "^\\d{5}(-\\d{4})?$",
-              "message": "Invalid ZIP code format",
-              "error_code": "zip_invalid_format"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "id": "preferences",
-      "name": "User Preferences",
-      "description": "User preference settings",
-      "type": "object",
-      "required": false,
-      "properties": [
-        {
-          "id": "theme",
-          "name": "Theme",
-          "type": "string",
-          "required": false,
-          "validation": [
-            {
-              "type": "enum",
-              "value": ["light", "dark", "system"],
-              "message": "Theme must be one of: light, dark, system",
-              "error_code": "theme_invalid_value"
-            }
-          ],
-          "default": "system"
-        },
-        {
-          "id": "notifications",
-          "name": "Notifications Enabled",
-          "type": "boolean",
-          "required": false,
-          "default": true
-        }
-      ]
-    }
-  ]
-}
