--- examples/workflow/string_workflow.yaml
@@ -1,31 +1,0 @@
-id: "string-workflow"
-name: "String Processing Workflow"
-description: "A workflow that demonstrates string processing operations"
-
-nodes:
-  - id: "input"
-    type: "string_processor"
-    provider: "string"
-    config:
-      operation: "uppercase"
-      input: "Hello, World!"
-
-  - id: "process"
-    type: "string_processor"
-    provider: "string"
-    config:
-      operation: "lowercase"
-      input: "${input.output}"
-
-  - id: "output"
-    type: "string_processor"
-    provider: "string"
-    config:
-      operation: "trim"
-      input: "${process.output}"
-
-edges:
-  - from: "input"
-    to: "process"
-  - from: "process"
-    to: "output"
