--- tests/logic_processor_test.go
@@ -1,97 +1,0 @@
-package tests
-
-import (
-	"context"
-	"testing"
-
-	"github.com/open-source-cloud/fuse/internal/workflow"
-	"github.com/open-source-cloud/fuse/pkg/logic"
-	"github.com/stretchr/testify/suite"
-)
-
-type LogicProcessorTestSuite struct {
-	suite.Suite
-	provider engine.NodeProvider
-}
-
-func (s *LogicProcessorTestSuite) SetupTest() {
-	s.provider = debug.NewLogicProcessorProvider()
-}
-
-func (s *LogicProcessorTestSuite) TestAnd() {
-	config := map[string]interface{}{
-		"type":   debug.NodeTypeAnd,
-		"values": []interface{}{true, true},
-	}
-
-	node, err := s.provider.CreateNode(config)
-	s.Require().NoError(err)
-	s.Require().NotNil(node)
-
-	result, err := node.Execute(context.TODO(), true)
-	s.Require().NoError(err)
-	s.Equal(true, result)
-
-	result, err = node.Execute(context.TODO(), false)
-	s.Require().NoError(err)
-	s.Equal(false, result)
-}
-
-func (s *LogicProcessorTestSuite) TestOr() {
-	config := map[string]interface{}{
-		"type":   debug.NodeTypeOr,
-		"values": []interface{}{false, true},
-	}
-
-	node, err := s.provider.CreateNode(config)
-	s.Require().NoError(err)
-	s.Require().NotNil(node)
-
-	result, err := node.Execute(context.TODO(), false)
-	s.Require().NoError(err)
-	s.Equal(true, result)
-
-	result, err = node.Execute(context.TODO(), true)
-	s.Require().NoError(err)
-	s.Equal(true, result)
-}
-
-func (s *LogicProcessorTestSuite) TestInvalidConfig() {
-	config := map[string]interface{}{
-		"invalid": "config",
-	}
-
-	node, err := s.provider.CreateNode(config)
-	s.Require().Error(err)
-	s.Nil(node)
-}
-
-func (s *LogicProcessorTestSuite) TestEmptyValues() {
-	config := map[string]interface{}{
-		"type":   debug.NodeTypeAnd,
-		"values": []interface{}{},
-	}
-
-	node, err := s.provider.CreateNode(config)
-	s.Require().Error(err)
-	s.Nil(node)
-}
-
-func (s *LogicProcessorTestSuite) TestInvalidValueType() {
-	config := map[string]interface{}{
-		"type":   debug.NodeTypeAnd,
-		"values": []interface{}{"not-a-boolean", "also-not-a-boolean"},
-	}
-
-	node, err := s.provider.CreateNode(config)
-	s.Require().NoError(err)
-	s.Require().NotNil(node)
-
-	result, err := node.Execute(context.TODO(), true)
-	s.Require().Error(err)
-	s.Nil(result)
-}
-
-func TestLogicProcessorSuite(t *testing.T) {
-	suite.Run(t, new(LogicProcessorTestSuite))
-}
